# back/central/shared/docs/auth/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Servicio REST para la gestión de reservas multi-restaurante.",
        "title": "Restaurant Reservation API",
        "termsOfService": "https://ejemplo.com/terminos",
        "contact": {
            "name": "Equipo de Backend",
            "email": "backend@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:3050",
    "basePath": "/api/v1",
    "paths": {
...
        "/roles/system": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Obtiene solo los roles del sistema (is_system = true)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Roles"
                ],
                "summary": "Obtener roles del sistema",
                "responses": {
                    "200": {
                        "description": "Roles del sistema obtenidos exitosamente",
                        "schema": {
                            "$ref": "#/definitions/response.RoleListResponse"
                        }
                    },
                    "401": {
                        "description": "Token de acceso requerido",
                        "schema": {
                            "$ref": "#/definitions/response.RoleErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "$ref": "#/definitions/response.RoleErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/users": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Obtiene la lista filtrada y paginada de usuarios del sistema con sus roles y businesses",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Obtener usuarios filtrados y paginados",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "default": 1,
                        "description": "Número de página",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "default": 10,
                        "description": "Tamaño de página",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filtrar por nombre (búsqueda parcial)",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filtrar por email (búsqueda parcial)",
                        "name": "email",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filtrar por teléfono (búsqueda parcial)",
                        "name": "phone",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filtrar por IDs de usuarios separados por comas (ej: 1,2,3)",
                        "name": "user_ids",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Filtrar por estado activo",
                        "name": "is_active",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filtrar por ID de rol",
                        "name": "role_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filtrar por ID de business",
                        "name": "business_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filtrar por fecha de creación (YYYY-MM-DD o YYYY-MM-DD,YYYY-MM-DD para rango)",
                        "name": "created_at",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "id",
                            "name",
                            "email",
                            "phone",
                            "is_active",
                            "created_at",
                            "updated_at"
                        ],
                        "type": "string",
                        "default": "created_at",
                        "description": "Campo para ordenar",
                        "name": "sort_by",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "asc",
                            "desc"
                        ],
                        "type": "string",
                        "default": "desc",
                        "description": "Orden de clasificación",
                        "name": "sort_order",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuarios obtenidos exitosamente",
                        "schema": {
                            "$ref": "#/definitions/response.UserListResponse"
                        }
                    },
                    "400": {
                        "description": "Parámetros de filtro inválidos",
                        "schema": {
                            "$ref": "#/definitions/response.UserErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Token de acceso requerido",
                        "schema": {
                            "$ref": "#/definitions/response.UserErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "$ref": "#/definitions/response.UserErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Crea un nuevo usuario.\nENVÍO: SOLO multipart/form-data (no JSON body)",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Crear usuario",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Nombre (2-100)",
                        "name": "name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Email válido",
                        "name": "email",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Teléfono (exactamente 10 dígitos)",
                        "name": "phone",
                        "in": "formData"
                    },
                    {
                        "type": "boolean",
                        "description": "¿Activo? (true/false)",
                        "name": "is_active",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "Imagen de avatar (sube a S3)",
                        "name": "avatarFile",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "IDs de negocios separados por comas. Ej: '16,21'",
                        "name": "business_ids",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Usuario creado exitosamente (incluye contraseña generada)",
                        "schema": {
                            "$ref": "#/definitions/response.UserCreatedResponse"
                        }
                    },
                    "400": {
                        "description": "Datos inválidos (mensajes claros de validación)",
                        "schema": {
                            "$ref": "#/definitions/response.UserErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Token de acceso requerido",
                        "schema": {
                            "$ref": "#/definitions/response.UserErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Email ya existe",
                        "schema": {
                            "$ref": "#/definitions/response.UserErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "$ref": "#/definitions/response.UserErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "response.BusinessRoleAssignmentResponse": {
            "type": "object",
            "properties": {
                "business_id": {
                    "type": "integer"
                },
                "business_name": {
                    "type": "string"
                },
                "role_id": {
                    "type": "integer"
                },
                "role_name": {
                    "type": "string"
                }
            }
        },
...
        "response.RoleErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Error interno del servidor"
                }
            }
        },
        "response.RoleListResponse": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer",
                    "example": 5
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.RoleResponse"
                    }
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "response.RoleResponse": {
            "type": "object",
            "properties": {
                "business_type_id": {
                    "type": "integer",
                    "example": 1
                },
                "business_type_name": {
                    "type": "string",
                    "example": "Propiedad Horizontal"
                },
                "code": {
                    "type": "string",
                    "example": "admin"
                },
                "description": {
                    "type": "string",
                    "example": "Rol de administrador del sistema"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "is_system": {
                    "type": "boolean",
                    "example": true
                },
                "level": {
                    "type": "integer",
                    "example": 2
                },
                "name": {
                    "type": "string",
                    "example": "Administrador"
                },
                "scope_code": {
                    "type": "string",
                    "example": "system"
                },
                "scope_id": {
                    "type": "integer",
                    "example": 1
                },
                "scope_name": {
                    "type": "string",
                    "example": "Sistema"
                }
            }
        },
...
        "response.UserCreatedResponse": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "response.UserErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "response.UserListResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.UserResponse"
                    }
                },
                "pagination": {
                    "$ref": "#/definitions/services_auth_users_internal_infra_primary_handlers_response.PaginationInfo"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "response.UserResponse": {
            "type": "object",
            "properties": {
                "avatar_url": {
                    "type": "string"
                },
                "business_role_assignments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.BusinessRoleAssignmentResponse"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_active": {
                    "type": "boolean"
                },
                "is_super_user": {
                    "type": "boolean"
                },
                "last_login_at": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "services_auth_users_internal_infra_primary_handlers_response.PaginationInfo": {
            "type": "object",
            "properties": {
                "current_page": {
                    "type": "integer"
                },
                "has_next": {
                    "type": "boolean"
                },
                "has_prev": {
                    "type": "boolean"
                },
                "last_page": {
                    "type": "integer"
                },
                "per_page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Token de business JWT con el prefijo **Bearer** (para todos los endpoints)",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "BusinessTokenAuth": {
            "description": "Token principal JWT con el prefijo **Bearer** (solo para /auth/business-token)",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}